Aha! There might be a time when part of a texture partialy visible becouse of occurence at left-up corner... and then what?! That state should be perfectly valid!

+++ include/graphics_core/graphics_engine.hpp
@@ -13,7 +13,7 @@ public:
     void begin_render(SDL_Color background_color);
     void end_render();
 
-    void draw(SDL_Texture *texture, size_t x, size_t y);
+    void draw(SDL_Texture *texture, int x, int y);
     void draw(SDL_Rect rectangle, SDL_Color color);
 
     SDL_Texture *render_text(TTF_Font *font, const std::string &text, SDL_Color color, size_t font_size);
